+++ apps/modernjs-ssr/nested-remote/serverroutebootstrap.js
@@ -1,0 +1,17 @@
+export const routes = [
+  {
+    path: '/',
+    children: [
+      {
+        _component: '@_modern_js_src/routes/page',
+        index: true,
+        id: 'page',
+        type: 'nested',
+      },
+    ],
+    isRoot: true,
+    _component: '@_modern_js_src/routes/layout',
+    id: 'layout',
+    type: 'nested',
+  },
+];