+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @module-federation/esbuild
 
+## 0.0.11
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.3
+
+## 0.0.10
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.2
+
 ## 0.0.9
 
 ### Patch Changes