+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,17 @@
 # [3.1.0](https://github.com/module-federation/core/compare/utils-3.0.2...utils-3.1.0) (2023-10-26)
 
+## 3.0.39
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.3
+
+## 3.0.38
+
+### Patch Changes
+
+- @module-federation/sdk@0.3.2
+
 ## 3.0.37
 
 ### Patch Changes